Add final files
