--- openvino-AI-project/README.md
@@ -1,29 +1,0 @@
-# openvino-AI-project
-openvino AI project를 위한 github (Human chasing mobility)
-
-
-# Project Gantt Chart
-
-```mermaid
-gantt
-    title Project Schedule
-    dateFormat  MM/DD
-    section Market Research
-    Pre-research        :active,   pre-research, 09/25, 1d
-    section AI Model
-    Human Detection     :active,   human-detection, 09/25, 3d
-    Hand Detection      :          hand-detection, 09/27, 3d
-    Hand Gesture Classification:   hand-gesture-classification, 09/30, 3d
-    Mono-eye Distance Measurement: mono-eye-distance, 09/25, 6d
-    Color Detection     :          color-detection, 09/25, 6d
-    Segmentation Model  :          segmentation, 09/25, 10d
-    Model Synthesis (Ensemble) :    model-synthesis, 09/30, 9d
-    Model Testing       :          model-testing, 09/27, 14d
-    section Hardware
-    Hardware Setup      :          hardware-setup, 10/08, 3d
-    Vehicle Algorithm Development:  vehicle-algorithm, 10/10, 7d
-    Embedded Software Development: embedded-software, 10/10, 7d
-    section Total Inspection
-    Debugging           :          debugging, 10/16, 6d
-    Testing             :          testing, 10/16, 7d
-    Rehearsal           :          rehearsal, 10/21, 2d
