--- openvino-AI-project/requirements.txt
@@ -1,172 +1,0 @@
-absl-py==2.1.0
-anyio==4.6.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==2.4.1
-astunparse==1.6.3
-async-lru==2.0.4
-attrs==24.2.0
-babel==2.16.0
-beautifulsoup4==4.12.3
-bleach==6.1.0
-certifi==2024.8.30
-cffi==1.17.1
-charset-normalizer==3.3.2
-comm==0.2.2
-contourpy==1.3.0
-cycler==0.12.1
-debugpy==1.8.6
-decorator==5.1.1
-defusedxml==0.7.1
-exceptiongroup==1.2.2
-executing==2.1.0
-fastjsonschema==2.20.0
-filelock==3.16.1
-flatbuffers==24.3.25
-fonttools==4.54.1
-fqdn==1.5.1
-fsspec==2024.9.0
-gast==0.6.0
-google-pasta==0.2.0
-grpcio==1.66.1
-h11==0.14.0
-h5py==3.11.0
-httpcore==1.0.5
-httpx==0.27.2
-idna==3.10
-ipykernel==6.29.5
-ipython==8.27.0
-ipywidgets==8.1.5
-isoduration==20.11.0
-jedi==0.19.1
-Jinja2==3.1.4
-json5==0.9.25
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2023.12.1
-jupyter==1.1.1
-jupyter_client==8.6.3
-jupyter-console==6.6.3
-jupyter_core==5.7.2
-jupyter-events==0.10.0
-jupyter-lsp==2.2.5
-jupyter_server==2.14.2
-jupyter_server_terminals==0.5.3
-jupyterlab==4.2.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
-kagglehub==0.3.0
-keras==3.5.0
-kiwisolver==1.4.7
-lab==8.2
-libclang==18.1.1
-Markdown==3.7
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-matplotlib==3.9.2
-matplotlib-inline==0.1.7
-mdurl==0.1.2
-mistune==3.0.2
-ml-dtypes==0.4.1
-mpmath==1.3.0
-namex==0.0.8
-nbclient==0.10.0
-nbconvert==7.16.4
-nbformat==5.10.4
-nest-asyncio==1.6.0
-networkx==3.1
-notebook==7.2.2
-notebook_shim==0.2.4
-numpy==1.26.4
-nvidia-cublas-cu12==12.1.3.1
-nvidia-cuda-cupti-cu12==12.1.105
-nvidia-cuda-nvrtc-cu12==12.1.105
-nvidia-cuda-runtime-cu12==12.1.105
-nvidia-cudnn-cu12==9.1.0.70
-nvidia-cufft-cu12==11.0.2.54
-nvidia-curand-cu12==10.3.2.106
-nvidia-cusolver-cu12==11.4.5.107
-nvidia-cusparse-cu12==12.1.0.106
-nvidia-nccl-cu12==2.20.5
-nvidia-nvjitlink-cu12==12.6.68
-nvidia-nvtx-cu12==12.1.105
-opencv-contrib-python==4.10.0.84
-opencv-python==4.10.0.84
-openvino==2024.4.0
-openvino-dev==2024.4.0
-openvino-telemetry==2024.1.0
-opt-einsum==3.3.0
-optree==0.12.1
-overrides==7.7.0
-packaging==24.1
-pandas==2.2.3
-pandocfilters==1.5.1
-parso==0.8.4
-pexpect==4.9.0
-pillow==10.4.0
-pip==22.0.2
-platformdirs==4.3.6
-prometheus_client==0.21.0
-prompt_toolkit==3.0.47
-protobuf==4.25.5
-psutil==6.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-py-cpuinfo==9.0.0
-pycparser==2.22
-Pygments==2.18.0
-pyparsing==3.1.4
-python-dateutil==2.9.0.post0
-python-json-logger==2.0.7
-pytz==2024.2
-PyYAML==6.0.2
-pyzmq==26.2.0
-referencing==0.35.1
-requests==2.32.3
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rich==13.8.1
-rpds-py==0.20.0
-scipy==1.14.1
-seaborn==0.13.2
-Send2Trash==1.8.3
-setuptools==59.6.0
-simplejson==3.19.3
-six==1.16.0
-sniffio==1.3.1
-soupsieve==2.6
-stack-data==0.6.3
-sympy==1.13.3
-tensorboard==2.17.1
-tensorboard-data-server==0.7.2
-tensorflow==2.17.0
-tensorflow-io-gcs-filesystem==0.37.1
-termcolor==2.4.0
-terminado==0.18.1
-tinycss2==1.3.0
-tomli==2.0.1
-torch==2.4.1
-torchaudio==2.4.1+cu124
-torchvision==0.19.1
-tornado==6.4.1
-tqdm==4.66.5
-traitlets==5.14.3
-triton==3.0.0
-txt2tags==3.9
-types-python-dateutil==2.9.0.20240906
-typing_extensions==4.12.2
-tzdata==2024.2
-ultralytics==8.2.100
-ultralytics-thop==2.0.8
-uri-template==1.3.0
-urllib3==2.2.3
-wcwidth==0.2.13
-webcolors==24.8.0
-webencodings==0.5.1
-websocket-client==1.8.0
-Werkzeug==3.0.4
-wheel==0.44.0
-widgetsnbextension==4.0.13
-wrapt==1.16.0
